requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
openai==1.3.7
pydantic==2.5.0
sqlalchemy==2.0.23
google-generativeai>=0.3.2

# Machine Learning
xgboost==2.0.3
scikit-learn==1.3.2
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.50

# Technical Analysis
pandas-ta
joblib
vaderSentiment
textblob

